+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: Relative path could be used for `Backup Path`
 - Fix: An error with `Only on change`, which was not working properly
 - Add: Option to create zip archive
+- Add: Option to specify the `Backup set name` if multiple backups are to be keep.
 
 ## v0.5.3 (2021-04-03)